Guess only possible answers when fewer than 3
When there are 1 or 2 remaining possible answers, it is optimal to
select guesses from among the answers, since the best worst-case
scenario is two turns whether we do so or not, but the algorithm can
waste a lot of time comparing all possible guesses to achieve the same
result.

We also noticed an error in benchmarks, namely the base and length
parameters were backwards in some tests. We have corrected this and
re-run benchmarks for all tagged commits where necessary.